requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
fonttools==4.54.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==3.0.1
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pi==0.1.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
